Add missing error check in resolve callback
If an error happens in libunbound during the processing of a resolve
request the callback will be called with err being set to a nonzero
value and result being NULL. Add a check to catch that error. This
requires slightly extending the data we pass to the callback ourselves
so the forwarder can still tell the parent process which query failed.