Add error check for libunbound init
Libunbound is only getting fully initialized during the first resolve
request. We do this by issuing a request for localhost during startup.
This can fail though if for example the configured trust anchor file
does not exist or the trust anchor is malformed. If that happens we
just exit pfresolved since further requests would fail anyways.