Add control socket and implement pfresolvectl
Adds a control process and control socket to pfresolved, allowing it to
be controlled from other programs. Also adds a new program called
'pfresolvectl' that utilizes this socket. It can be used to change the
log level of pfresolved, reload the current configuration, or trigger
pfresolved to write the hints file if configured.

The implementation is based on 'control.c' from iked(8) and on
ikectl(8).